The Theme Utility

Auto generate theme for user.

get_theme(
  data,
  theme_name = "npg",
  threshold = 10,
  mode = "fill",
  reverse = False,
  palette = NULL,
  alpha = 1
)

Arguments

data
the data that is mapping with color.

theme_name
theme name Currently there is 19 available options:

"npg"

"aaas"

"nejm"

"lancet"

"jama"

"jco"

"ucscgb"

"d3"

"locuszoom"

"igv"

"cosmic"

"uchicago"

"startrek"

"tron"

"futurama"

"rickandmorty"

"simpsons"

"gsea"

"material"

See the Scientific Journal and Sci-Fi Themed Color Palettes for ggplot2 for details.

threshold
Integer. It is a continuous variable when the number of unique element is greater than threshold

reverse
Logical. Should the order of the colors be reversed?

palette
Palette type, only when data is a continuous variable.

alpha
Transparency level, a real number in (0, 1].

Author

Jingcheng Yang <yjcyxky@163.com>

Examples

library("BioVisReportR")
theme <- get_theme(c(1:16), theme_name = 'npg')

#> Error in get_theme(c(1:16), theme_name = "npg"): The theme npg is not match data type.